requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0
jinja2==3.1.4
aiofiles==23.2.1
